Moving code and removing DBCStore code
