+++ src/game/Server/DBCStores.cpp
@@ -16,797 +16,4 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "Server/DBCStores.h"
-#include "Policies/Singleton.h"
-#include "Log.h"
-#include "ProgressBar.h"
-#include "Util.h"
-#include "Globals/Locales.h"
-#include "Globals/SharedDefines.h"
-#include "Server/SQLStorages.h"
 
-#include "DBCfmt.h"
-
-#include <map>
-
-// typedef std::map<uint16, uint32> AreaFlagByAreaID;
-// typedef std::map<uint32, uint32> AreaFlagByMapID;
-//
-// struct WMOAreaTableTripple
-// {
-//     WMOAreaTableTripple(int32 r, int32 a, int32 g) : groupId(g), rootId(r), adtId(a)
-//     {
-//     }
-//
-//     bool operator <(const WMOAreaTableTripple& b) const
-//     {
-//         return memcmp(this, &b, sizeof(WMOAreaTableTripple)) < 0;
-//     }
-//
-//     // ordered by entropy; that way memcmp will have a minimal medium runtime
-//     int32 groupId;
-//     int32 rootId;
-//     int32 adtId;
-// };
-//
-// typedef std::map<WMOAreaTableTripple, std::vector<WMOAreaTableEntry const*>> WMOAreaInfoByTripple;
-//
-// DBCStorage <AreaTableEntry> sDBCAreaTable(AreaTableEntryfmt);
-// static AreaFlagByAreaID sAreaFlagByAreaID;
-// static AreaFlagByMapID  sAreaFlagByMapID;                   // for instances without generated *.map files
-//
-// static WMOAreaInfoByTripple sWMOAreaInfoByTripple;
-
-//DBCStorage <AreaTriggerEntry> sDBCAreaTrigger(AreaTriggerEntryfmt);
-//DBCStorage <AuctionHouseEntry> sDBCAuctionHouse(AuctionHouseEntryfmt);
-//DBCStorage <BankBagSlotPricesEntry> sDBCBankBagSlotPrices(BankBagSlotPricesEntryfmt);
-//DBCStorage <BattlemasterListEntry> sDBCBattlemasterList(BattlemasterListEntryfmt);
-//DBCStorage <CharStartOutfitEntry> sDBCCharStartOutfit(CharStartOutfitEntryfmt);
-//DBCStorage <CharTitlesEntry> sDBCCharTitles(CharTitlesEntryfmt);
-//DBCStorage <ChatChannelsEntry> sDBCChatChannels(ChatChannelsEntryfmt);
-//DBCStorage <ChrClassesEntry> sDBCChrClasses(ChrClassesEntryfmt);
-//DBCStorage <ChrRacesEntry> sDBCChrRaces(ChrRacesEntryfmt);
-//DBCStorage <CinematicCameraEntry> sDBCCinematicCamera(CinematicCameraEntryfmt);
-//DBCStorage <CinematicSequencesEntry> sDBCCinematicSequences(CinematicSequencesEntryfmt);
-//DBCStorage <CreatureDisplayInfoEntry> sDBCCreatureDisplayInfo(CreatureDisplayInfofmt);
-//DBCStorage <CreatureDisplayInfoExtraEntry> sDBCCreatureDisplayInfoExtra(CreatureDisplayInfoExtrafmt);
-//DBCStorage <CreatureFamilyEntry> sDBCCreatureFamily(CreatureFamilyfmt);
-//DBCStorage <CreatureModelDataEntry> sDBCCreatureModelData(CreatureModelDatafmt);
-//DBCStorage <CreatureSpellDataEntry> sDBCCreatureSpellData(CreatureSpellDatafmt);
-//DBCStorage <CreatureTypeEntry> sDBCCreatureType(CreatureTypefmt);
-
-//DBCStorage <DurabilityQualityEntry> sDBCDurabilityCosts(DurabilityQualityfmt);
-//DBCStorage <DurabilityCostsEntry> sDBCDurabilityQuality(DurabilityCostsfmt);
-
-//DBCStorage <EmotesEntry> sDBCEmotes(EmotesEntryfmt);
-//DBCStorage <EmotesTextEntry> sDBCEmotesText(EmotesTextEntryfmt);
-
-//DBCStorage <FactionEntry> sFactionStore(FactionEntryfmt);
-//DBCStorage <FactionTemplateEntry> sDBCFactionTemplate(FactionTemplateEntryfmt);
-
-//DBCStorage <GameObjectDisplayInfoEntry> sDBCGameObjectDisplayInfo(GameObjectDisplayInfofmt);
-//DBCStorage <GemPropertiesEntry> sDBCGemProperties(GemPropertiesEntryfmt);
-
-//DBCStorage <GMSurveyCurrentSurveyEntry> sDBCGMSurveyCurrentSurvey(GMSurveyCurrentSurveyfmt);
-//DBCStorage <GMSurveyQuestionsEntry> sDBCGMSurveyQuestions(GMSurveyQuestionsfmt);
-//DBCStorage <GMSurveyEntry> sDBCGMSurveySurveys(GMSurveySurveysfmt);
-//DBCStorage <GMTicketCategoryEntry> sDBCGMTicketCategory(GMTicketCategoryfmt);
-
-//DBCStorage <GtCombatRatingsEntry>         sDBCgtCombatRatings(GtCombatRatingsfmt);
-//DBCStorage <GtChanceToMeleeCritBaseEntry> sDBCgtChanceToMeleeCritBase(GtChanceToMeleeCritBasefmt);
-//DBCStorage <GtChanceToMeleeCritEntry>     sDBCgtChanceToMeleeCrit(GtChanceToMeleeCritfmt);
-//DBCStorage <GtChanceToSpellCritBaseEntry> sDBCgtChanceToSpellCritBase(GtChanceToSpellCritBasefmt);
-//DBCStorage <GtChanceToSpellCritEntry>     sDBCgtChanceToSpellCrit(GtChanceToSpellCritfmt);
-//DBCStorage <GtOCTRegenHPEntry>            sDBCgtOCTRegenHP(GtOCTRegenHPfmt);
-//DBCStorage <GtNPCManaCostScalerEntry>     sDBCgtNPCManaCostScaler(GtNPCManaCostScalerfmt);
-// DBCStorage <GtOCTRegenMPEntry>            sGtOCTRegenMPStore(GtOCTRegenMPfmt);  -- not used currently
-//DBCStorage <GtRegenHPPerSptEntry>         sDBCgtRegenHPPerSpt(GtRegenHPPerSptfmt);
-//DBCStorage <GtRegenMPPerSptEntry>         sDBCgtRegenMPPerSpt(GtRegenMPPerSptfmt);
-
-//DBCStorage <ItemEntry>                    sDBCItem(Itemfmt);
-//DBCStorage <ItemBagFamilyEntry>           sDBCItemBagFamily(ItemBagFamilyfmt);
-//DBCStorage <ItemClassEntry>               sDBCItemClass(ItemClassfmt);
-// DBCStorage <ItemCondExtCostsEntry> sItemCondExtCostsStore(ItemCondExtCostsEntryfmt);
-// DBCStorage <ItemDisplayInfoEntry> sItemDisplayInfoStore(ItemDisplayTemplateEntryfmt); -- not used currently
-//DBCStorage <ItemExtendedCostEntry> sDBCItemExtendedCost(ItemExtendedCostEntryfmt);
-//DBCStorage <ItemRandomPropertiesEntry> sDBCItemRandomProperties(ItemRandomPropertiesfmt);
-//DBCStorage <ItemRandomSuffixEntry> sDBCItemRandomSuffix(ItemRandomSuffixfmt);
-//DBCStorage <ItemSetEntry> sDBCItemSet(ItemSetEntryfmt);
-//DBCStorage <LiquidTypeEntry> sDBCLiquidType(LiquidTypefmt);
-//DBCStorage <LockEntry> sDBCLock(LockEntryfmt);
-
-//DBCStorage <MailTemplateEntry> sDBCMailTemplate(MailTemplateEntryfmt);
-//DBCStorage <MapEntry> sDBCMap(MapEntryfmt);
-
-//DBCStorage <QuestSortEntry> sDBCQuestSort(QuestSortEntryfmt);
-
-//DBCStorage <RandomPropertiesPointsEntry> sDBCRandPropPoints(RandomPropertiesPointsfmt);
-
-//DBCStorage <SkillLineEntry> sDBCSkillLine(SkillLinefmt);
-//DBCStorage <SkillLineAbilityEntry> sDBCSkillLineAbility(SkillLineAbilityfmt);
-//DBCStorage <SkillRaceClassInfoEntry> sDBCSkillRaceClassInfo(SkillRaceClassInfofmt);
-//DBCStorage <SkillTiersEntry> sDBCSkillTiers(SkillTiersfmt);
-
-//DBCStorage <SoundEntriesEntry> sDBCSoundEntries(SoundEntriesfmt);
-
-//DBCStorage <SpellItemEnchantmentEntry> sDBCSpellItemEnchantment(SpellItemEnchantmentfmt);
-//DBCStorage <SpellItemEnchantmentConditionEntry> sDBCSpellItemEnchantmentCondition(SpellItemEnchantmentConditionfmt);
-//DBCStorage <SpellVisualEntry> sDBCSpellVisual(SpellVisualfmt);
-SpellCategoryStore sSpellCategoryStore;
-ItemSpellCategoryStore sItemSpellCategoryStore;
-PetFamilySpellsStore sPetFamilySpellsStore;
-
-//DBCStorage <SpellCastTimesEntry> sDBCSpellCastTimes(SpellCastTimefmt);
-//DBCStorage <SpellDurationEntry> sDBCSpellDuration(SpellDurationfmt);
-//DBCStorage <SpellFocusObjectEntry> sDBCSpellFocusObject(SpellFocusObjectfmt);
-//DBCStorage <SpellRadiusEntry> sDBCSpellRadius(SpellRadiusfmt);
-//DBCStorage <SpellRangeEntry> sDBCSpellRange(SpellRangefmt);
-//DBCStorage <SpellShapeshiftFormEntry> sDBCSpellShapeshiftForm(SpellShapeshiftfmt);
-//DBCStorage <StableSlotPricesEntry> sDBCStableSlotPrices(StableSlotPricesfmt);
-//DBCStorage <SummonPropertiesEntry> sDBCSummonProperties(SummonPropertiesfmt);
-//DBCStorage <TalentEntry> sDBCTalent(TalentEntryfmt);
-TalentSpellPosMap sTalentSpellPosMap;
-//DBCStorage <TalentTabEntry> sDBCTalentTab(TalentTabEntryfmt);
-
-// store absolute bit position for first rank for talent inspect
-typedef std::map<uint32, uint32> TalentInspectMap;
-static TalentInspectMap sTalentPosInInspect;
-static TalentInspectMap sTalentTabSizeInInspect;
-static uint32 sTalentTabPages[12/*MAX_CLASSES*/][3];
-
-//DBCStorage <TaxiNodesEntry> sDBCTaxiNodes(TaxiNodesEntryfmt);
-TaxiMask sTaxiNodesMask;
-
-// DBC used only for initialization sTaxiPathSetBySource at startup.
-TaxiPathSetBySource sTaxiPathSetBySource;
-//DBCStorage <TaxiPathEntry> sDBCTaxiPath(TaxiPathEntryfmt);
-
-// DBC store data but sTaxiPathNodesByPath used for fast access to entries (it's not owner pointed data).
-TaxiPathNodesByPath sTaxiPathNodesByPath;
-//static DBCStorage <TaxiPathNodeEntry> sDBCTaxiPathNode(TaxiPathNodeEntryfmt);
-
-//DBCStorage <TransportAnimationEntry> sDBCTransportAnimation(TransportAnimationfmt);
-
-//DBCStorage <TotemCategoryEntry> sDBCTotemCategory(TotemCategoryEntryfmt);
-//DBCStorage <WMOAreaTableEntry>  sDBCWMOAreaTable(WMOAreaTableEntryfmt);
-//DBCStorage <WorldMapAreaEntry>  sDBCWorldMapArea(WorldMapAreaEntryfmt);
-// DBCStorage <WorldMapOverlayEntry> sWorldMapOverlayStore(WorldMapOverlayEntryfmt);
-
-typedef std::list<std::string> StoreProblemList;
-
-bool IsAcceptableClientBuild(uint32 build)
-{
-    int accepted_versions[] = EXPECTED_MANGOSD_CLIENT_BUILD;
-    for (int i = 0; accepted_versions[i]; ++i)
-        if (int(build) == accepted_versions[i])
-            return true;
-
-    return false;
-}
-
-std::string AcceptableClientBuildsListStr()
-{
-    std::ostringstream data;
-    int accepted_versions[] = EXPECTED_MANGOSD_CLIENT_BUILD;
-    for (int i = 0; accepted_versions[i]; ++i)
-        data << accepted_versions[i] << " ";
-    return data.str();
-}
-
-static bool LoadDBC_assert_print(uint32 fsize, uint32 rsize, const std::string& filename)
-{
-    sLog.outError("Size of '%s' setted by format string (%u) not equal size of C++ structure (%u).", filename.c_str(), fsize, rsize);
-
-    // ASSERT must fail after function call
-    return false;
-}
-
-template<class T>
-inline void LoadDBC(uint32& availableDbcLocales, BarGoLink& bar, StoreProblemList& errlist, DBCStorage<T>& storage, const std::string& dbc_path, const std::string& filename)
-{
-    // compatibility format and C++ structure sizes
-    MANGOS_ASSERT(DBCFileLoader::GetFormatRecordSize(storage.GetFormat()) == sizeof(T) || LoadDBC_assert_print(DBCFileLoader::GetFormatRecordSize(storage.GetFormat()), sizeof(T), filename));
-
-    std::string dbc_filename = dbc_path + filename;
-    if (storage.Load(dbc_filename.c_str()))
-    {
-        bar.step();
-        for (uint8 i = 0; fullLocaleNameList[i].name; ++i)
-        {
-            if (!(availableDbcLocales & (1 << i)))
-                continue;
-
-            std::string dbc_filename_loc = dbc_path + fullLocaleNameList[i].name + "/" + filename;
-            if (!storage.LoadStringsFrom(dbc_filename_loc.c_str()))
-                availableDbcLocales &= ~(1 << i);           // mark as not available for speedup next checks
-        }
-    }
-    else
-    {
-        // sort problematic dbc to (1) non compatible and (2) nonexistent
-        FILE* f = fopen(dbc_filename.c_str(), "rb");
-        if (f)
-        {
-            char buf[100];
-            snprintf(buf, 100, " (exist, but have %u fields instead " SIZEFMTD ") Wrong client version DBC file?", storage.GetFieldCount(), strlen(storage.GetFormat()));
-            errlist.push_back(dbc_filename + buf);
-            fclose(f);
-        }
-        else
-            errlist.push_back(dbc_filename);
-    }
-}
-
-void LoadDBCStores(const std::string& dataPath)
-{
-    std::string dbcPath = dataPath + "dbc/";
-
-    const uint32 DBCFilesCount = 66;
-
-    BarGoLink bar(DBCFilesCount);
-
-    StoreProblemList bad_dbc_files;
-
-    // bitmask for index of fullLocaleNameList
-    uint32 availableDbcLocales = 0xFFFFFFFF;
-
-//     LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCAreaTable,                dbcPath, "AreaTable.dbc");
-//
-//     // must be after sAreaStore loading
-//     for (uint32 i = 0; i < sDBCAreaTable.GetNumRows(); ++i)    // areaflag numbered from 0
-//     {
-//         if (AreaTableEntry const* area = sDBCAreaTable.LookupEntry(i))
-//         {
-//             // fill AreaId->DBC records
-//             sAreaFlagByAreaID.insert(AreaFlagByAreaID::value_type(uint16(area->ID), area->exploreFlag));
-//
-//             // fill MapId->DBC records ( skip sub zones and continents )
-//             if (area->zone == 0 && area->mapid != 0 && area->mapid != 1 && area->mapid != 530)
-//                 sAreaFlagByMapID.insert(AreaFlagByMapID::value_type(area->mapid, area->exploreFlag));
-//         }
-//     }
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCAreaTrigger,         dbcPath, "AreaTrigger.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCAuctionHouse,        dbcPath, "AuctionHouse.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCBankBagSlotPrices,   dbcPath, "BankBagSlotPrices.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCBattlemasterList,    dbcPath, "BattlemasterList.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCCharStartOutfit,     dbcPath, "CharStartOutfit.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCCharTitles,          dbcPath, "CharTitles.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCChatChannels,        dbcPath, "ChatChannels.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCChrClasses,          dbcPath, "ChrClasses.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDBCChrRaces,            dbcPath, "ChrRaces.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCinematicCameraStore,     dbcPath, "CinematicCamera.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCinematicSequencesStore,  dbcPath, "CinematicSequences.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureDisplayInfoStore, dbcPath, "CreatureDisplayInfo.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureDisplayInfoExtraStore, dbcPath, "CreatureDisplayInfoExtra.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureFamilyStore,      dbcPath, "CreatureFamily.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureModelDataStore,   dbcPath, "CreatureModelData.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureSpellDataStore,   dbcPath, "CreatureSpellData.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sCreatureTypeStore,        dbcPath, "CreatureType.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDurabilityCostsStore,     dbcPath, "DurabilityCosts.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sDurabilityQualityStore,   dbcPath, "DurabilityQuality.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sEmotesStore,              dbcPath, "Emotes.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sEmotesTextStore,          dbcPath, "EmotesText.dbc");
-    // LoadDBC(availableDbcLocales, bar, bad_dbc_files, sFactionStore,             dbcPath, "Faction.dbc");
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sFactionTemplateStore,     dbcPath, "FactionTemplate.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGameObjectDisplayInfoStore, dbcPath, "GameObjectDisplayInfo.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGemPropertiesStore,       dbcPath, "GemProperties.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGMSurveyCurrentSurveyStore,  dbcPath, "GMSurveyCurrentSurvey.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGMSurveyQuestionsStore,  dbcPath, "GMSurveyQuestions.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGMSurveySurveysStore,  dbcPath, "GMSurveySurveys.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGMTicketCategoryStore, dbcPath, "GMTicketCategory.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtCombatRatingsStore,     dbcPath, "gtCombatRatings.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtChanceToMeleeCritBaseStore, dbcPath, "gtChanceToMeleeCritBase.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtChanceToMeleeCritStore, dbcPath, "gtChanceToMeleeCrit.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtChanceToSpellCritBaseStore, dbcPath, "gtChanceToSpellCritBase.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtChanceToSpellCritStore, dbcPath, "gtChanceToSpellCrit.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtOCTRegenHPStore,        dbcPath, "gtOCTRegenHP.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtNPCManaCostScalerStore, dbcPath, "gtNPCManaCostScaler.dbc");
-//    // LoadDBC(availableDbcLocales,bar,bad_dbc_files,sGtOCTRegenMPStore,        dbcPath,"gtOCTRegenMP.dbc");       -- not used currently
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtRegenHPPerSptStore,     dbcPath, "gtRegenHPPerSpt.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sGtRegenMPPerSptStore,     dbcPath, "gtRegenMPPerSpt.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemStore,                dbcPath, "Item.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemBagFamilyStore,       dbcPath, "ItemBagFamily.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemClassStore,           dbcPath, "ItemClass.dbc");
-//    // LoadDBC(availableDbcLocales,bar,bad_dbc_files,sItemDisplayInfoStore,     dbcPath,"ItemDisplayInfo.dbc");     -- not used currently
-//    // LoadDBC(availableDbcLocales,bar,bad_dbc_files,sItemCondExtCostsStore,    dbcPath,"ItemCondExtCosts.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemExtendedCostStore,    dbcPath, "ItemExtendedCost.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemRandomPropertiesStore, dbcPath, "ItemRandomProperties.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemRandomSuffixStore,    dbcPath, "ItemRandomSuffix.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sItemSetStore,             dbcPath, "ItemSet.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sLiquidTypeStore,          dbcPath, "LiquidType.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sLockStore,                dbcPath, "Lock.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sMailTemplateStore,        dbcPath, "MailTemplate.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sMapStore,                 dbcPath, "Map.dbc");
-    //{
-    //    // repairs entry for netherstorm - should be moved to SQL
-    //    MapEntry const* mEntry = sDBCMap.LookupEntry(550);
-    //    MapEntry* tempestKeepMap = new MapEntry(*mEntry);
-    //    tempestKeepMap->ghost_entrance_map = 530;
-    //    sDBCMap.EraseEntry(550);
-    //    sDBCMap.InsertEntry(tempestKeepMap, 550);
-    //}
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sQuestSortStore,           dbcPath, "QuestSort.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sRandomPropertiesPointsStore, dbcPath, "RandPropPoints.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSkillLineStore,           dbcPath, "SkillLine.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSkillLineAbilityStore,    dbcPath, "SkillLineAbility.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSkillRaceClassInfoStore,  dbcPath, "SkillRaceClassInfo.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSkillTiersStore,          dbcPath, "SkillTiers.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSoundEntriesStore,        dbcPath, "SoundEntries.dbc");
-
-    for (auto skillLine : sDBCSkillLineAbility)
-    {
-        SpellEntry const* spellInfo = sSpellTemplate.LookupEntry(skillLine->spellId);
-        if (spellInfo && (spellInfo->Attributes & (SPELL_ATTR_ABILITY | SPELL_ATTR_PASSIVE | SPELL_ATTR_HIDDEN_CLIENTSIDE | SPELL_ATTR_HIDE_IN_COMBAT_LOG)) == (SPELL_ATTR_ABILITY | SPELL_ATTR_PASSIVE | SPELL_ATTR_HIDDEN_CLIENTSIDE | SPELL_ATTR_HIDE_IN_COMBAT_LOG))
-        {
-            for (auto cFamily : sDBCCreatureFamily)
-            {
-                if (skillLine->skillId != cFamily->skillLine[0] && skillLine->skillId != cFamily->skillLine[1])
-                    continue;
-
-                sPetFamilySpellsStore[cFamily->ID].insert(spellInfo->Id);
-            }
-        }
-    }
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellCastTimesStore,      dbcPath, "SpellCastTimes.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellDurationStore,       dbcPath, "SpellDuration.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellFocusObjectStore,    dbcPath, "SpellFocusObject.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellItemEnchantmentStore, dbcPath, "SpellItemEnchantment.dbc");
-    //for (uint32 i = 0; i < sSpellItemEnchantmentStore.GetNumRows(); ++i)
-    //{
-    //    SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(i);
-    //    if (!enchantEntry)
-    //        continue;
-    //    for (uint32 k = 0; k < 3; ++k)
-    //    {
-    //        if (enchantEntry->spellid[k])
-    //            if (!sSpellTemplate.LookupEntry(enchantEntry->spellid[k]))
-    //                sLog.outErrorDb("Spell ID %u found in spell item enchant %u does not exist.", enchantEntry->spellid[k], i);
-    //    }
-    //}
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellItemEnchantmentConditionStore, dbcPath, "SpellItemEnchantmentCondition.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellRadiusStore,         dbcPath, "SpellRadius.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellRangeStore,          dbcPath, "SpellRange.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellShapeshiftFormStore, dbcPath, "SpellShapeshiftForm.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSpellVisualStore,         dbcPath, "SpellVisual.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sStableSlotPricesStore,    dbcPath, "StableSlotPrices.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sSummonPropertiesStore,    dbcPath, "SummonProperties.dbc");
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTalentStore,              dbcPath, "Talent.dbc");
-
-    // create talent spells set
-    for (auto talentInfo : sDBCTalent)
-    {
-        for (int j = 0; j < 5; ++j)
-            if (talentInfo->RankID[j])
-                sTalentSpellPosMap[talentInfo->RankID[j]] = TalentSpellPos(talentInfo->TalentID, j);
-    }
-
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTalentTabStore,           dbcPath, "TalentTab.dbc");
-
-    // prepare fast data access to bit pos of talent ranks for use at inspecting
-    {
-        // fill table by amount of talent ranks and fill sTalentTabBitSizeInInspect
-        // store in with (row,col,talent)->size key for correct sorting by (row,col)
-        typedef std::map<uint32, uint32> TalentBitSize;
-        TalentBitSize sTalentBitSize;
-        for (auto talentInfo : sDBCTalent)
-        {
-            TalentTabEntry const* talentTabInfo = sDBCTalentTab.LookupEntry(talentInfo->TalentTab);
-            if (!talentTabInfo)
-                continue;
-
-            // find talent rank
-            uint32 curtalent_maxrank = 0;
-            for (uint32 k = 5; k > 0; --k)
-            {
-                if (talentInfo->RankID[k - 1])
-                {
-                    curtalent_maxrank = k;
-                    break;
-                }
-            }
-
-            sTalentBitSize[(talentInfo->Row << 24) + (talentInfo->Col << 16) + talentInfo->TalentID] = curtalent_maxrank;
-            sTalentTabSizeInInspect[talentInfo->TalentTab] += curtalent_maxrank;
-        }
-
-        // now have all max ranks (and then bit amount used for store talent ranks in inspect)
-        for (auto talentTabInfo : sDBCTalentTab)
-        {
-            // prevent memory corruption; otherwise cls will become 12 below
-            if ((talentTabInfo->ClassMask & CLASSMASK_ALL_PLAYABLE) == 0)
-                continue;
-
-            // store class talent tab pages
-            uint32 cls = 1;
-            for (uint32 m = 1; !(m & talentTabInfo->ClassMask) && cls < MAX_CLASSES; m <<= 1, ++cls) {}
-
-            sTalentTabPages[cls][talentTabInfo->tabpage] = talentTabInfo->TalentTabID;
-
-            // add total amount bits for first rank starting from talent tab first talent rank pos.
-            uint32 pos = 0;
-            for (TalentBitSize::iterator itr = sTalentBitSize.begin(); itr != sTalentBitSize.end(); ++itr)
-            {
-                uint32 talentId = itr->first & 0xFFFF;
-                TalentEntry const* talentInfo = sDBCTalent.LookupEntry(talentId);
-                if (!talentInfo)
-                    continue;
-
-                if (talentInfo->TalentTab != talentTabInfo->TalentTabID)
-                    continue;
-
-                sTalentPosInInspect[talentId] = pos;
-                pos += itr->second;
-            }
-        }
-    }
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTaxiNodesStore,           dbcPath, "TaxiNodes.dbc");
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTaxiPathStore,            dbcPath, "TaxiPath.dbc");
-    for (auto entry : sDBCTaxiPath)
-        sTaxiPathSetBySource[entry->from][entry->to] = TaxiPathBySourceAndDestination(entry->ID, entry->price);
-    uint32 pathCount = sDBCTaxiPath.GetMaxEntry();
-
-    //## TaxiPathNode.dbc ## Loaded only for initialization different structures
- //   LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTaxiPathNodeStore,        dbcPath, "TaxiPathNode.dbc");
-    // Calculate path nodes count
-    std::vector<uint32> pathLength;
-    pathLength.resize(pathCount);                           // 0 and some other indexes not used
-    for (auto entry : sDBCTaxiPathNode)
-    {
-         if (pathLength[entry->path] < entry->index + 1)
-            pathLength[entry->path] = entry->index + 1;
-    }
-    // Set path length
-    sTaxiPathNodesByPath.resize(pathCount);                 // 0 and some other indexes not used
-    for (uint32 i = 1; i < sTaxiPathNodesByPath.size(); ++i)
-        sTaxiPathNodesByPath[i].resize(pathLength[i]);
-    // fill data (pointers to sTaxiPathNodeStore elements
-    for (auto entry : sDBCTaxiPathNode)
-        sTaxiPathNodesByPath[entry->path].set(entry->index, entry);
-
-    // Initialize global taxinodes mask
-    // include existing nodes that have at least single not spell base (scripted) path
-    {
-        std::set<uint32> spellPaths;
-        for (auto sInfo : sSpellTemplate)
-            for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
-                if (sInfo->Effect[j] == 123 /*SPELL_EFFECT_SEND_TAXI*/)
-                    spellPaths.insert(sInfo->EffectMiscValue[j]);
-
-        memset(sTaxiNodesMask, 0, sizeof(sTaxiNodesMask));
-        for (auto node : sDBCTaxiNodes)
-        {
-            TaxiPathSetBySource::const_iterator src_i = sTaxiPathSetBySource.find(node->ID);
-            if (src_i != sTaxiPathSetBySource.end() && !src_i->second.empty())
-            {
-                bool ok = false;
-                for (const auto& dest_i : src_i->second)
-                {
-                    // not spell path
-                    if (spellPaths.find(dest_i.second.ID) == spellPaths.end())
-                    {
-                        ok = true;
-                        break;
-                    }
-                }
-
-                if (!ok)
-                    continue;
-            }
-
-            // valid taxi network node
-            uint8  field   = (uint8)((node->ID - 1) / 32);
-            uint32 submask = 1 << ((node->ID - 1) % 32);
-            sTaxiNodesMask[field] |= submask;
-        }
-    }
-
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTransportAnimationStore,  dbcPath, "TransportAnimation.dbc");
-//
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sTotemCategoryStore,       dbcPath, "TotemCategory.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sWorldMapAreaStore,        dbcPath, "WorldMapArea.dbc");
-//    LoadDBC(availableDbcLocales, bar, bad_dbc_files, sWMOAreaTableStore,        dbcPath, "WMOAreaTable.dbc");
-    for (auto entry : sDBCWMOAreaTable)
-    {
-        sWMOAreaInfoByTripple[WMOAreaTableTripple(entry->rootId, entry->adtId, entry->groupId)].push_back(entry);
-    }
-    // LoadDBC(availableDbcLocales,bar,bad_dbc_files,sWorldMapOverlayStore,     dbcPath,"WorldMapOverlay.dbc");
-    // LoadDBC(availableDbcLocales, bar, bad_dbc_files, sWorldSafeLocsStore,       dbcPath, "WorldSafeLocs.dbc");
-
-    // error checks
-    if (bad_dbc_files.size() >= DBCFilesCount)
-    {
-        sLog.outError("\nIncorrect DataDir value in mangosd.conf or ALL required *.dbc files (%d) not found by path: %sdbc", DBCFilesCount, dataPath.c_str());
-        Log::WaitBeforeContinueIfNeed();
-        exit(1);
-    }
-    if (!bad_dbc_files.empty())
-    {
-        std::string str;
-        for (auto& bad_dbc_file : bad_dbc_files)
-            str += bad_dbc_file + "\n";
-
-        sLog.outError("\nSome required *.dbc files (%u from %d) not found or not compatible:\n%s", (uint32)bad_dbc_files.size(), DBCFilesCount, str.c_str());
-        Log::WaitBeforeContinueIfNeed();
-        exit(1);
-    }
-
-    // Check loaded DBC files proper version
-    if (!sDBCSkillLineAbility.LookupEntry(17514) ||
-            !sDBCMap.LookupEntry(598)                ||
-            !sDBCGemProperties.LookupEntry(1127)     ||
-            !sDBCItemExtendedCost.LookupEntry(2425)  ||
-            !sDBCCharTitles.LookupEntry(71)          ||
-            !sDBCAreaTable.LookupEntry(1768))
-    {
-        sLog.outError("\nYou have _outdated_ DBC files. Please re-extract DBC files for one from client build: %s", AcceptableClientBuildsListStr().c_str());
-        Log::WaitBeforeContinueIfNeed();
-        exit(1);
-    }
-
-    sLog.outString(">> Initialized %d data stores", DBCFilesCount);
-    sLog.outString();
-}
-
-char const* GetPetName(uint32 petfamily, uint32 dbclang)
-{
-    if (!petfamily)
-        return nullptr;
-    CreatureFamilyEntry const* pet_family = sDBCCreatureFamily.LookupEntry(petfamily);
-    if (!pet_family)
-        return nullptr;
-    return pet_family->Name[dbclang] ? pet_family->Name[dbclang] : nullptr;
-}
-
-TalentSpellPos const* GetTalentSpellPos(uint32 spellId)
-{
-    TalentSpellPosMap::const_iterator itr = sTalentSpellPosMap.find(spellId);
-    if (itr == sTalentSpellPosMap.end())
-        return nullptr;
-
-    return &itr->second;
-}
-
-uint32 GetTalentSpellCost(TalentSpellPos const* pos)
-{
-    if (pos)
-        return pos->rank + 1;
-
-    return 0;
-}
-
-uint32 GetTalentSpellCost(uint32 spellId)
-{
-    return GetTalentSpellCost(GetTalentSpellPos(spellId));
-}
-
-
-
-uint32 GetVirtualMapForMapAndZone(uint32 mapid, uint32 zoneId)
-{
-    if (mapid != 530)                                       // speed for most cases
-        return mapid;
-
-    if (WorldMapAreaEntry const* wma = sDBCWorldMapArea.LookupEntry(zoneId))
-        return wma->virtual_map_id >= 0 ? wma->virtual_map_id : wma->map_id;
-
-    return mapid;
-}
-
-ContentLevels GetContentLevelsForMapAndZone(uint32 mapid, uint32 zoneId)
-{
-    mapid = GetVirtualMapForMapAndZone(mapid, zoneId);
-    if (mapid < 2)
-        return CONTENT_1_60;
-
-    MapEntry const* mapEntry = sDBCMap.LookupEntry(mapid);
-    if (!mapEntry)
-        return CONTENT_1_60;
-
-    switch (mapEntry->Expansion())
-    {
-        default: return CONTENT_1_60;
-        case 1:  return CONTENT_61_70;
-    }
-}
-
-ChatChannelsEntry const* GetChatChannelsEntryFor(const std::string& name, uint32 channel_id/* = 0*/)
-{
-    std::wstring wname;
-
-    Utf8toWStr(name, wname);
-
-    if (!channel_id && wname.empty())
-        return nullptr;
-
-    // not sorted, numbering index from 0
-    for (auto entry : sDBCChatChannels)
-    {
-        std::wstring wpattern;
-
-        // try to match by name first, avoid creating custom channels with same name
-        if (!wname.empty())
-        {
-            for (uint32 i = 0; i < MAX_LOCALE; ++i)
-            {
-                Utf8toWStr(entry->pattern[i], wpattern);
-
-                if (wpattern.empty())
-                    continue;
-
-                size_t argpos = wpattern.find(L"%s");
-
-                // formatting arg present: strip and attempt partial match
-                if (argpos != std::wstring::npos)
-                {
-                    wpattern.replace(argpos, 2, L"");
-
-                    if (wname.find(wpattern) != std::wstring::npos)
-                        return entry;
-                }
-                // attempt full match
-                else if (wname.compare(wpattern) == 0)
-                    return entry;
-            }
-        }
-
-        // name still not found, but channel id is provided: possibly no dbc data for client locale
-        if (channel_id && channel_id == entry->ChannelID)
-            return entry;
-    }
-    return nullptr;
-}
-
-bool IsTotemCategoryCompatiableWith(uint32 itemTotemCategoryId, uint32 requiredTotemCategoryId)
-{
-    if (requiredTotemCategoryId == 0)
-        return true;
-    if (itemTotemCategoryId == 0)
-        return false;
-
-    TotemCategoryEntry const* itemEntry = sDBCTotemCategory.LookupEntry(itemTotemCategoryId);
-    if (!itemEntry)
-        return false;
-    TotemCategoryEntry const* reqEntry = sDBCTotemCategory.LookupEntry(requiredTotemCategoryId);
-    if (!reqEntry)
-        return false;
-
-    if (itemEntry->categoryType != reqEntry->categoryType)
-        return false;
-
-    return (itemEntry->categoryMask & reqEntry->categoryMask) == reqEntry->categoryMask;
-}
-
-bool MapCoordinateVsZoneCheck(float x, float y, uint32 mapid, uint32 zone)
-{
-    WorldMapAreaEntry const* WmaEntry = sDBCWorldMapArea.LookupEntry(zone);
-    if (WmaEntry && mapid == WmaEntry->map_id &&
-            x >= WmaEntry->x2 && x <= WmaEntry->x1 && y >= WmaEntry->y2 && y <= WmaEntry->y1)
-        return true;
-    return false;
-}
-
-bool Zone2MapCoordinates(float& x, float& y, uint32 zone)
-{
-    WorldMapAreaEntry const* maEntry = sDBCWorldMapArea.LookupEntry(zone);
-
-    // if not listed then map coordinates (instance)
-    if (!maEntry || maEntry->x2 == maEntry->x1 || maEntry->y2 == maEntry->y1)
-        return false;
-
-    std::swap(x, y);                                        // at client map coords swapped
-    x = x * ((maEntry->x2 - maEntry->x1) / 100) + maEntry->x1;
-    y = y * ((maEntry->y2 - maEntry->y1) / 100) + maEntry->y1; // client y coord from top to down
-
-    return true;
-}
-
-bool Map2ZoneCoordinates(float& x, float& y, uint32 zone)
-{
-    WorldMapAreaEntry const* maEntry = sDBCWorldMapArea.LookupEntry(zone);
-
-    // if not listed then map coordinates (instance)
-    if (!maEntry || maEntry->x2 == maEntry->x1 || maEntry->y2 == maEntry->y1)
-        return false;
-
-    x = (x - maEntry->x1) / ((maEntry->x2 - maEntry->x1) / 100);
-    y = (y - maEntry->y1) / ((maEntry->y2 - maEntry->y1) / 100); // client y coord from top to down
-    std::swap(x, y);                                        // client have map coords swapped
-
-    return true;
-}
-
-uint32 GetTalentInspectBitPosInTab(uint32 talentId)
-{
-    TalentInspectMap::const_iterator itr = sTalentPosInInspect.find(talentId);
-    if (itr == sTalentPosInInspect.end())
-        return 0;
-
-    return itr->second;
-}
-
-uint32 GetTalentTabInspectBitSize(uint32 talentTabId)
-{
-    TalentInspectMap::const_iterator itr = sTalentTabSizeInInspect.find(talentTabId);
-    if (itr == sTalentTabSizeInInspect.end())
-        return 0;
-
-    return itr->second;
-}
-
-uint32 const* GetTalentTabPages(uint32 cls)
-{
-    return sTalentTabPages[cls];
-}
-
-bool IsPointInAreaTriggerZone(AreaTriggerEntry const* atEntry, uint32 mapid, float x, float y, float z, float delta)
-{
-    if (mapid != atEntry->mapid)
-        return false;
-
-    if (atEntry->radius > 0)
-    {
-        // if we have radius check it
-        float dist2 = (x - atEntry->x) * (x - atEntry->x) + (y - atEntry->y) * (y - atEntry->y) + (z - atEntry->z) * (z - atEntry->z);
-        if (dist2 > (atEntry->radius + delta) * (atEntry->radius + delta))
-            return false;
-    }
-    else
-    {
-        // we have only extent
-
-        // rotate the players position instead of rotating the whole cube, that way we can make a simplified
-        // is-in-cube check and we have to calculate only one point instead of 4
-
-        // 2PI = 360, keep in mind that ingame orientation is counter-clockwise
-        double rotation = 2 * M_PI - atEntry->box_orientation;
-        double sinVal = sin(rotation);
-        double cosVal = cos(rotation);
-
-        float playerBoxDistX = x - atEntry->x;
-        float playerBoxDistY = y - atEntry->y;
-
-        float rotPlayerX = float(atEntry->x + playerBoxDistX * cosVal - playerBoxDistY * sinVal);
-        float rotPlayerY = float(atEntry->y + playerBoxDistY * cosVal + playerBoxDistX * sinVal);
-
-        // box edges are parallel to coordiante axis, so we can treat every dimension independently :D
-        float dz = z - atEntry->z;
-        float dx = rotPlayerX - atEntry->x;
-        float dy = rotPlayerY - atEntry->y;
-        if ((fabs(dx) > atEntry->box_x / 2 + delta) ||
-                (fabs(dy) > atEntry->box_y / 2 + delta) ||
-                (fabs(dz) > atEntry->box_z / 2 + delta))
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-uint32 GetCreatureModelRace(uint32 model_id)
-{
-    CreatureDisplayInfoEntry const* displayEntry = sDBCCreatureDisplayInfo.LookupEntry(model_id);
-    if (!displayEntry)
-        return 0;
-    CreatureDisplayInfoExtraEntry const* extraEntry = sDBCCreatureDisplayInfoExtra.LookupEntry(displayEntry->ExtendedDisplayInfoID);
-    return extraEntry ? extraEntry->Race : 0;
-}
-
-// script support functions
-//DBCStorage <SoundEntriesEntry>  const* GetSoundEntriesStore()   { return &sDBCSoundEntries;   }
-//DBCStorage <SpellRangeEntry>    const* GetSpellRangeStore()     { return &sDBCSpellRange;     }
-// DBCStorage <FactionEntry>       const* GetFactionStore()        { return &sFactionStore;        }
-//DBCStorage <ItemEntry>          const* GetItemDisplayStore()    { return &sDBCItem;           }
-//DBCStorage <CreatureDisplayInfoEntry> const* GetCreatureDisplayStore() { return &sDBCCreatureDisplayInfo; }
-//DBCStorage <EmotesEntry>        const* GetEmotesStore()         { return &sDBCEmotes;         }
-//DBCStorage <EmotesTextEntry>    const* GetEmotesTextStore()     { return &sDBCEmotesText;     }
