+++ src/game/Server/DBCStores.h
@@ -19,151 +19,4 @@
 #ifndef MANGOS_DBCSTORES_H
 #define MANGOS_DBCSTORES_H
 
-#include "Common.h"
-#include "Database/DBCStore.h"
-#include "Server/DBCStructure.h"
-
-#include <list>
-#include <vector>
-
-bool IsAcceptableClientBuild(uint32 build);
-std::string AcceptableClientBuildsListStr();
-
-char const* GetPetName(uint32 petfamily, uint32 dbclang);
-uint32 GetTalentSpellCost(uint32 spellId);
-uint32 GetTalentSpellCost(TalentSpellPos const* pos);
-TalentSpellPos const* GetTalentSpellPos(uint32 spellId);
-
-int32 GetAreaFlagByAreaID(uint32 area_id);                  // -1 if not found
-uint32 GetAreaFlagByMapId(uint32 mapid);
-
-std::vector<WMOAreaTableEntry const*>& GetWMOAreaTableEntriesByTripple(int32 rootid, int32 adtid, int32 groupid);
-
-AreaTableEntry const* GetAreaEntryByAreaID(uint32 area_id);
-AreaTableEntry const* GetAreaEntryByAreaFlagAndMap(uint32 area_flag, uint32 map_id);
-
-uint32 GetVirtualMapForMapAndZone(uint32 mapid, uint32 zoneId);
-
-enum ContentLevels
-{
-    CONTENT_1_60 = 0,
-    CONTENT_61_70
-};
-ContentLevels GetContentLevelsForMapAndZone(uint32 mapid, uint32 zoneId);
-
-ChatChannelsEntry const* GetChatChannelsEntryFor(const std::string& name, uint32 channel_id = 0);
-
-bool IsTotemCategoryCompatiableWith(uint32 itemTotemCategoryId, uint32 requiredTotemCategoryId);
-
-bool MapCoordinateVsZoneCheck(float x, float y, uint32 mapid, uint32 zone);
-bool Zone2MapCoordinates(float& x, float& y, uint32 zone);
-bool Map2ZoneCoordinates(float& x, float& y, uint32 zone);
-
-uint32 GetTalentInspectBitPosInTab(uint32 talentId);
-uint32 GetTalentTabInspectBitSize(uint32 talentTabId);
-uint32 const* /*[3]*/ GetTalentTabPages(uint32 cls);
-
-bool IsPointInAreaTriggerZone(AreaTriggerEntry const* atEntry, uint32 mapid, float x, float y, float z, float delta = 0.0f);
-
-uint32 GetCreatureModelRace(uint32 model_id);
-
-//extern DBCStorage <AreaTableEntry>               sDBCAreaTable;// recommend access using functions
-//extern DBCStorage <AreaTriggerEntry>             sDBCAreaTrigger;
-//extern DBCStorage <AuctionHouseEntry>            sDBCAuctionHouse;
-//extern DBCStorage <BankBagSlotPricesEntry>       sDBCBankBagSlotPrices;
-//extern DBCStorage <BattlemasterListEntry>        sDBCBattlemasterList;
-// extern DBCStorage <ChatChannelsEntry>           sChatChannelsStore; -- accessed using function, no usable index
-//extern DBCStorage <CharStartOutfitEntry>         sDBCCharStartOutfit;
-//extern DBCStorage <CharTitlesEntry>              sDBCCharTitles;
-//extern DBCStorage <ChatChannelsEntry>            sDBCChatChannels;
-//extern DBCStorage <ChrClassesEntry>              sDBCChrClasses;
-//extern DBCStorage <ChrRacesEntry>                sDBCChrRaces;
-//extern DBCStorage <CinematicCameraEntry>         sDBCCinematicCamera;
-//extern DBCStorage <CinematicSequencesEntry>      sCinematicSequencesStore;
-//extern DBCStorage <CreatureDisplayInfoEntry>     sCreatureDisplayInfoStore;
-//extern DBCStorage <CreatureDisplayInfoExtraEntry>sCreatureDisplayInfoExtraStore;
-//extern DBCStorage <CreatureFamilyEntry>          sCreatureFamilyStore;
-//extern DBCStorage <CreatureModelDataEntry>       sCreatureModelDataStore;
-//extern DBCStorage <CreatureSpellDataEntry>       sCreatureSpellDataStore;
-//extern DBCStorage <CreatureTypeEntry>            sCreatureTypeStore;
-//extern DBCStorage <DurabilityCostsEntry>         sDurabilityCostsStore;
-//extern DBCStorage <DurabilityQualityEntry>       sDurabilityQualityStore;
-//extern DBCStorage <EmotesEntry>                  sEmotesStore;
-//extern DBCStorage <EmotesTextEntry>              sEmotesTextStore;
-////extern DBCStorage <FactionEntry>                 sFactionStore;
-//extern DBCStorage <FactionTemplateEntry>         sFactionTemplateStore;
-//extern DBCStorage <GameObjectDisplayInfoEntry>   sGameObjectDisplayInfoStore;
-//extern DBCStorage <GemPropertiesEntry>           sGemPropertiesStore;
-//
-//extern DBCStorage <GMSurveyCurrentSurveyEntry>   sGMSurveyCurrentSurveyStore;
-//extern DBCStorage <GMSurveyQuestionsEntry>       sGMSurveyQuestionsStore;
-//extern DBCStorage <GMSurveyEntry>                sGMSurveySurveysStore;
-//extern DBCStorage <GMTicketCategoryEntry>        sGMTicketCategoryStore;
-//
-//extern DBCStorage <GtCombatRatingsEntry>         sGtCombatRatingsStore;
-//extern DBCStorage <GtChanceToMeleeCritBaseEntry> sGtChanceToMeleeCritBaseStore;
-//extern DBCStorage <GtChanceToMeleeCritEntry>     sGtChanceToMeleeCritStore;
-//extern DBCStorage <GtChanceToSpellCritBaseEntry> sGtChanceToSpellCritBaseStore;
-//extern DBCStorage <GtChanceToSpellCritEntry>     sGtChanceToSpellCritStore;
-//extern DBCStorage <GtOCTRegenHPEntry>            sGtOCTRegenHPStore;
-//extern DBCStorage <GtNPCManaCostScalerEntry>     sGtNPCManaCostScalerStore;
-//// extern DBCStorage <GtOCTRegenMPEntry>            sGtOCTRegenMPStore; -- not used currently
-//extern DBCStorage <GtRegenHPPerSptEntry>         sGtRegenHPPerSptStore;
-//extern DBCStorage <GtRegenMPPerSptEntry>         sGtRegenMPPerSptStore;
-//extern DBCStorage <ItemEntry>                    sItemStore;
-//extern DBCStorage <ItemBagFamilyEntry>           sItemBagFamilyStore;
-//extern DBCStorage <ItemClassEntry>               sItemClassStore;
-//// extern DBCStorage <ItemDisplayInfoEntry>      sItemDisplayInfoStore; -- not used currently
-//extern DBCStorage <ItemExtendedCostEntry>        sItemExtendedCostStore;
-//extern DBCStorage <ItemRandomPropertiesEntry>    sItemRandomPropertiesStore;
-//extern DBCStorage <ItemRandomSuffixEntry>        sItemRandomSuffixStore;
-//extern DBCStorage <ItemSetEntry>                 sItemSetStore;
-//extern DBCStorage <LiquidTypeEntry>              sLiquidTypeStore;
-//extern DBCStorage <LockEntry>                    sLockStore;
-//extern DBCStorage <MailTemplateEntry>            sMailTemplateStore;
-//extern DBCStorage <MapEntry>                     sMapStore;
-//extern DBCStorage <QuestSortEntry>               sQuestSortStore;
-//extern DBCStorage <RandomPropertiesPointsEntry>  sRandomPropertiesPointsStore;
-//extern DBCStorage <SkillLineEntry>               sSkillLineStore;
-//extern DBCStorage <SkillLineAbilityEntry>        sSkillLineAbilityStore;
-//extern DBCStorage <SkillRaceClassInfoEntry>      sSkillRaceClassInfoStore;
-//extern DBCStorage <SkillTiersEntry>              sSkillTiersStore;
-//extern DBCStorage <SoundEntriesEntry>            sSoundEntriesStore;
-//extern DBCStorage <SpellCastTimesEntry>          sSpellCastTimesStore;
-//extern DBCStorage <SpellDurationEntry>           sSpellDurationStore;
-//extern DBCStorage <SpellFocusObjectEntry>        sSpellFocusObjectStore;
-//extern DBCStorage <SpellItemEnchantmentEntry>    sSpellItemEnchantmentStore;
-//extern DBCStorage <SpellItemEnchantmentConditionEntry> sSpellItemEnchantmentConditionStore;
-extern SpellCategoryStore                        sSpellCategoryStore;
-extern ItemSpellCategoryStore                    sItemSpellCategoryStore;
-extern PetFamilySpellsStore                      sPetFamilySpellsStore;
-//extern DBCStorage <SpellRadiusEntry>             sSpellRadiusStore;
-//extern DBCStorage <SpellRangeEntry>              sSpellRangeStore;
-//extern DBCStorage <SpellShapeshiftFormEntry>     sSpellShapeshiftFormStore;
-//extern DBCStorage <SpellVisualEntry>             sSpellVisualStore;
-//extern DBCStorage <StableSlotPricesEntry>        sStableSlotPricesStore;
-//extern DBCStorage <SummonPropertiesEntry>        sSummonPropertiesStore;
-//extern DBCStorage <TalentEntry>                  sTalentStore;
-//extern DBCStorage <TalentTabEntry>               sTalentTabStore;
-//extern DBCStorage <TaxiNodesEntry>               sTaxiNodesStore;
-//extern DBCStorage <TaxiPathEntry>                sTaxiPathStore;
-extern TaxiMask                                  sTaxiNodesMask;
-extern TaxiPathSetBySource                       sTaxiPathSetBySource;
-extern TaxiPathNodesByPath                       sTaxiPathNodesByPath;
-//extern DBCStorage <TransportAnimationEntry>      sTransportAnimationStore;
-//extern DBCStorage <TotemCategoryEntry>           sTotemCategoryStore;
-//extern DBCStorage <WMOAreaTableEntry>            sWMOAreaTableStore;
-//// extern DBCStorage <WorldMapAreaEntry>           sWorldMapAreaStore; -- use Zone2MapCoordinates and Map2ZoneCoordinates
-//// extern DBCStorage <WorldMapOverlayEntry>         sWorldMapOverlayStore; -- not used currently
-//
-void LoadDBCStores(const std::string& dataPath);
-//
-//// script support functions
-//DBCStorage <SoundEntriesEntry>          const* GetSoundEntriesStore();
-//DBCStorage <SpellRangeEntry>            const* GetSpellRangeStore();
-//DBCStorage <FactionEntry>               const* GetFactionStore();
-//DBCStorage <ItemEntry>                  const* GetItemDisplayStore();
-//DBCStorage <CreatureDisplayInfoEntry>   const* GetCreatureDisplayStore();
-//DBCStorage <EmotesEntry>                const* GetEmotesStore();
-//DBCStorage <EmotesTextEntry>            const* GetEmotesTextStore();
 #endif
