+++ src/game/World/World.cpp
@@ -895,7 +895,7 @@ void World::SetInitialWorldSettings()
     ///- Load the DBC files
     sLog.outString("Initialize DBC data stores...");
     LoadDBCTables();
-    LoadDBCStores(m_dataPath);
+    //LoadDBCStores(m_dataPath);
     DetectDBCLang();
     sObjectMgr.SetDbc2StorageLocaleIndex(GetDefaultDbcLocale());    // Get once for all the locale index of DBC language (console/broadcasts)
 
@@ -2623,4 +2623,23 @@ void World::GeneratePacketMetrics()
     meas_players.add_field("mage", std::to_string(GetOnlineClassPlayers(CLASS_MAGE)));
     meas_players.add_field("warlock", std::to_string(GetOnlineClassPlayers(CLASS_WARLOCK)));
     meas_players.add_field("druid", std::to_string(GetOnlineClassPlayers(CLASS_DRUID)));
+}
+
+bool World::IsAcceptableClientBuild(uint32 build)
+{
+    int accepted_versions[] = EXPECTED_MANGOSD_CLIENT_BUILD;
+    for (int i = 0; accepted_versions[i]; ++i)
+        if (int(build) == accepted_versions[i])
+            return true;
+
+    return false;
+}
+
+std::string World::AcceptableClientBuildsListStr()
+{
+    std::ostringstream data;
+    int accepted_versions[] = EXPECTED_MANGOSD_CLIENT_BUILD;
+    for (int i = 0; accepted_versions[i]; ++i)
+        data << accepted_versions[i] << " ";
+    return data.str();
 }
