+++ src/mangosd/Master.cpp
@@ -143,7 +143,7 @@ int Master::Run()
 
     // set realmbuilds depend on mangosd expected builds, and set server online
     {
-        std::string builds = AcceptableClientBuildsListStr();
+        std::string builds = World::AcceptableClientBuildsListStr();
         LoginDatabase.escape_string(builds);
         LoginDatabase.DirectPExecute("UPDATE realmlist SET realmflags = realmflags & ~(%u), population = 0, realmbuilds = '%s'  WHERE id = '%u'", REALM_FLAG_OFFLINE, builds.c_str(), realmID);
     }
